+++ packages/cxalio-studio-tools/cx_tools/app/__init__.py
@@ -1,14 +1,20 @@
 __all__ = [
     "ConfigManager",
     "CxHighlighter",
+    "IAppComponent",
+    "IAppContext",
     "IAppEnvironment",
     "IApplication",
+    "run_async",
     "SafeError",
     "TextFileOpener",
     "try_open_text_file",
 ]
 
+from .async_runner import run_async
 from .config_manager import ConfigManager
+from .iappcomponent import IAppComponent
+from .iappcontext import IAppContext
 from .iappenv import CxHighlighter, IAppEnvironment
 from .iapplication import IApplication
 from .progress_task_agent import ProgressTaskAgent