+++ packages/cxalio-studio-tools/ffpretty/__init__.py
@@ -1,14 +1,27 @@
 __version__ = "0.9.0"
 
+import sys
+
+from .appcontext import AppContext
 from .application import FFPrettyApp
+from .appenv import appenv
 
 
 def run():
     from rich.traceback import install
+    from cx_tools.app import SafeError
 
     install(show_locals=False, word_wrap=True, suppress=["rich"])
 
-    with FFPrettyApp() as app:
-        result = app.run()
-        return 0 if result else -1
+    with appenv:
+        try:
+            context = AppContext.from_arguments(sys.argv[1:])
+        except SafeError as e:
+            appenv.say(f"[{e.style}]{e}[/]")
+            return
+        with FFPrettyApp(
+            appenv=appenv, context=context, progress=appenv.progress
+        ) as app:
+            result = app.run()
+            return 0 if result else -1
     return None