+++ packages/cxalio-studio-tools/media_scout/__init__.py
@@ -1,11 +1,16 @@
 __version__ = "0.8.6"
 
-from .application import Application
+from .application import MediaScoutApp
+from .appenv import AppEnv
+from .arg_parser import AppContext
 
 
 def run():
     from rich.traceback import install
 
     install(show_locals=True)
-    with Application() as app:
-        app.run()
+    context = AppContext.load()
+    appenv = AppEnv()
+    with appenv:
+        with MediaScoutApp(appenv=appenv, context=context) as app:
+            app.run()